+++ react/typings/graphql/mutations/addItem.gql.d.ts
@@ -1,0 +1,15 @@
+declare module '*/addItem.gql' {
+  import { DocumentNode } from 'graphql'
+
+  export interface Args {
+    subscriptionId: string
+    item: {
+      id: string
+      quantity: number
+    }
+  }
+
+  const value: DocumentNode
+  export default value
+}
+ˆ