+++ react/typings/vtex.styleguide.d.ts
@@ -16,6 +16,12 @@ declare module 'vtex.styleguide' {
   export const ActionMenu
   export const NumericStepper
   export const CheckboxGroup
+  export const ButtonWithIcon
+  export const IconPlus
+  export const Modal
+  export const InputSearch
+  export const ButtonPlain
+  export const Tooltip
   export interface ShowToastArgs {
     message: string
   }