+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Add product feature on the subscription details page.
+
 ## [3.3.0] - 2020-08-04
 
 ### Changed